requirements.txt
fastapi[all]==0.109.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.17.0
google-cloud-pubsub==2.19.0
twilio==8.11.0
sendgrid==6.11.0
uvicorn[standard]==0.27.0
structlog==24.1.0
pytest==7.4.4
httpx==0.26.0
python-multipart==0.0.6
